[Refactor]: 메시지쓰기버튼 common
